Add files via upload
The sieve of Eratosthenes is an ancient algorithm for finding all prime numbers up to any given limit. By iteratively marking as composite (i.e., not prime) the multiples of each prime, starting with the first prime number. The multiples of a given prime are generated as a sequence of numbers starting from that prime, with constant difference between them that is equal to that prime. This is the sieve's key distinction from using trial division to sequentially test each candidate number for divisibility by each prime. Once all the multiples of each discovered prime have been marked as composites, the remaining unmarked numbers are primes.
I set the procedure stop at the mid-point of the series.